перенос filter в board-presenter
